requirements.txt
contractions==0.1.73
en-core-web-sm==3.7.1
fasttext==0.9.2
fasttext-wheel==0.9.2
fqdn==1.5.1
isoduration==20.11.0
jsonpointer==2.4
natsort==8.4.0
nltk==3.8.1
notebook==7.2.1
openpyxl==3.1.5
pip-chill==1.0.3
pycld2==0.42
pyenchant==3.2.2
pyhelpers==2.0.0rc1
pyldavis==3.4.1
pyspellchecker==0.8.1
pytest-cov==5.0.0
rapidfuzz==3.9.4
sacremoses==0.1.1
seaborn==0.13.2
sphinx-copybutton==0.5.2
sphinx-rtd-theme==2.0.0
uri-template==1.3.0
vadersentiment==3.3.2
webcolors==1.13
